% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_diverg.R
\name{barplot.diverg}
\alias{barplot.diverg}
\title{barplot.diverg}
\usage{
\method{barplot}{diverg}(data = data.v1.2,
  var1 = data.v1.2$La.situation.d.emploi, var2 = data.v1.2$dep_T,
  interactive.vars = TRUE, marginvar = 2, vardroite = 3,
  vargauche = 1, coldroite = couleurs$bg1, colgauche = couleurs$rp1,
  varplus = c(2), colplus = couleurs$grayStable, varna = 4,
  colna = couleurs$grayNA, naou = 1, add.na.axis = TRUE, BYNA = 10,
  colo.vec = colo, XLIM = c(-100, 20), BYX = 20,
  YLABS = "Ensemble", par.left.1 = 8, par.left.2 = 2,
  Titre = "La situation de l'emploi...",
  Comment = "Enquète réalisée en septembre 2017 par la MRIE Rhône-Alpes",
  line.titre = -1, line.comment = -2, XLAB = "\%", bornmin = 100,
  barswidth = 5, YLIM = c(1, 5))
}
\arguments{
\item{data}{data.v1.2 ; le data.frame}

\item{var1}{data.v1.2$La.situation.d.emploi ; le data.frame ind?x?}

\item{var2}{data.v1.2$dep_T ; le data.frame ind?x?. var2 repr?sente les strates, les cat?gories.}

\item{interactive.vars}{TRUE ; faut-il une aide pour indexer les variables?}

\item{marginvar}{2 ; la marge sur laquelle sont calcul?s les pourcentages}

\item{vardroite}{3 ; l'indexe (num?rique) de la (ou des) variables ? droite}

\item{vargauche}{1 ; l'indexe (num?rique) de la (ou des) variables ? gauche}

\item{coldroite}{couleurs$bg1 ; couleurs ? droite (vecteur de character)}

\item{colgauche}{couleurs$rp1 ; couleurs ? gauche (vecteur de character)}

\item{varplus}{c(2) ; une variable en plus? Dans le cas d'une r?ponse interm?diaire (ni droite ni gauche)}

\item{colplus}{couleurs$grayStable ; couleur pour la ou les variable en plus}

\item{varna}{4 ; l'indexe (num?rique) de la (ou des) variables d?signant les non r?ponses, NA...}

\item{colna}{couleurs$grayNA ; couleur pour la ou les variable NA}

\item{naou}{1 ; NA sur le graph 1 ou 2 (le cas ?ch?ant)}

\item{add.na.axis}{TRUE ; faut-il un axe sp?cifique aux NA}

\item{BYNA}{10 ; interval minimal pour l'axe des NA}

\item{colo.vec}{colo ; DEPRECATED}

\item{XLIM}{c(-100,20) ; XLIM pour le graph 1. Pensez ? une borne inf?rieure suffisante pour les NA}

\item{BYX}{20 ; interval minmal pour l'axe X principal (graph 1 et 2)}

\item{YLABS}{colnames(dat) ; nom des cat?gories (var2)}

\item{par.left.1}{8 ; mar ? gauche pour le graph 1 (laisser de la place pour YLABS)}

\item{par.left.2}{2 , mar ? gauche pour le graph 2}

\item{Titre}{"La situation de l'emploi..."}

\item{Comment}{"Enqu?te r?alis?e en septembre 2017 par la MRIE Rh?ne-Alpes"}

\item{line.titre}{-1}

\item{line.comment}{-2}

\item{XLAB}{"%"}
}
\value{
la fonction ne retourne rien du tout.
}
\description{
barplot.diverg permet de repr?senter une variable factor issue d'une enqu?te de mani?re divergente
}
\author{
Elie
}
