% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_reponse_multi.R
\name{barplot.reponse.multi}
\alias{barplot.reponse.multi}
\title{barplot.reponse.multi permet de représenter conjointement deux variables qualitatives}
\usage{
\method{barplot}{reponse.multi}(tab.eff = tab.profil.urb,
  tab.prop = prop.tab.profil.urb, var1, var2, margin.prop = 1,
  mar = c(7, 10, 4, 2), xlim = c(0, 0.4), BYX = 0.1, XLAB = "\%",
  colo = c(couleurs$rur.green, couleurs$urb.blue), wrap.len.labels,
  legend.labels = c("Rural", "Urbain"), x.legend = 0.15,
  y.legend = 10, axe.title.lab = c("Type:", "Effectifs \\n totaux:",
  "Part des répondants \\n concernés:"), axe.title.at = c(-0.075,
  -0.05, 0.2), lecture.lab, add.test = TRUE,
  test = fish.test.profil.urb, box.test.y.ajust = 0, y.amplitude = 4,
  box.test.x.ajust = 0.1, x.amplitude = 0.11,
  test.spe = "Test de Fisher (petits effectifs)", print.tab = FALSE)
}
\arguments{
\item{tab.eff}{=tab.profil.urb, table des effectifs, peut être NULL}

\item{tab.prop}{=prop.tab.profil.urb,table des proportions, peut être NULL}

\item{var1}{si tab.eff et tab.prop sont NULL}

\item{var2}{idem}

\item{margin.prop}{=1, marge sur laquelle calculer les proportions}

\item{mar}{=c(7, 10, 4, 2),par(mar)}

\item{xlim}{=c(0, 0.4),}

\item{XLAB}{="%"}

\item{colo}{=c(couleurs$rur.green, couleurs$urb.blue),}

\item{wrap.len.labels}{=" ", paramètre len pour wrapper les labels}

\item{legend.labels}{=c("Rural", "Urbain"),}

\item{x.legend}{=0.15,}

\item{y.legend}{=10,}

\item{axe.title.lab}{= c("Type:", "Effectifs n totaux:", "Part des répondants  concernés:"),}

\item{axe.title.at}{=c(-0.075, -0.05, 0.2),}

\item{lecture.lab}{=}

\item{add.test}{=TRUE, y a-t-il un test d'indépendance (fisher, Khi2...)}

\item{test}{=fish.test.profil.urb,  variable résultat de test}

\item{box.test.y.ajust}{=0, décalage entre la test box et la légende sur l'axe y}

\item{y.amplitude}{=4 , amplitude du déclage sur y}

\item{box.test.x.ajust}{=0.1, décalage entre la test box et la légende sur l'axe x}

\item{x.amplitude}{=0.11, amplitude du déclage sur y}

\item{test.spe}{="Test de Fisher (petits effectifs)"}

\item{BYX=0.1}{interval minimal pour l'échelle des x}
}
\value{
some value
}
\description{
barplot.reponse.multi permet de représenter conjointement deux variables qualitatives
}
