% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_barchart}
\alias{create_barchart}
\title{Create a bar chart
Create a Bar Chart}
\usage{
create_barchart(data, x_var, y_var)
}
\arguments{
\item{data}{A data frame.}

\item{x_var}{Name of the x-axis column.}

\item{y_var}{Name of the y-axis column.}
}
\value{
A ggplot2 bar chart object.
}
\description{
Create a bar chart
Create a Bar Chart
}
