% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_histogram}
\alias{create_histogram}
\title{Create a histogram}
\usage{
create_histogram(data, x_var, bins = 30)
}
\arguments{
\item{data}{A data frame.}

\item{x_var}{A string name of the column for x-axis (numeric).}

\item{bins}{Number of bins (default = 30).}
}
\value{
A ggplot2 histogram object.
}
\description{
Create a histogram
}
