% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_scatterplot}
\alias{create_scatterplot}
\title{Create a scatter plot
Create a Scatter Plot}
\usage{
create_scatterplot(data, x_var, y_var)
}
\arguments{
\item{data}{A data frame.}

\item{x_var}{X-axis variable name (string).}

\item{y_var}{Y-axis variable name (string).}
}
\value{
A ggplot2 scatter plot object.
}
\description{
Create a scatter plot
Create a Scatter Plot
}
