% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ala}
\alias{data_ala}
\title{RT for three levels of Valence}
\format{
A data frame with 273 rows and 3 variables:
\describe{
  \item{Participant}{An identifier for the participant}
  \item{Valence}{"Sad", "Neutral", or "Happy"}
  \item{RT}{Response Time in seconds}
}
}
\source{
Ala
}
\usage{
data_ala
}
\description{
A data set with the RT scores of three groups.
}
\keyword{datasets}
