% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ldt}
\alias{data_ldt}
\title{Lexical Decision Task data}
\format{
A data frame:
\describe{
  \item{id}{An identifier for the participant}
  \item{age}{age}
  \item{language}{mono or bi: "monolingual", "bilingual"}
  \item{acc}{Mean Accuracy for the current participant in the current condition}
  \item{rt}{Mean RT for the current participant in the current condition}
}
}
\source{
introdataviz package from PsyTeachR
}
\usage{
data_ldt
}
\description{
A data set with condition mean RT and accuracy for the lexical decision task
}
\keyword{datasets}
