% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mc}
\alias{data_mc}
\title{RT and Accuracy for Videos}
\format{
A data frame:
\describe{
  \item{pp}{An identifier for the participant}
  \item{Age}{Age}
  \item{Valence}{"Negative", "Neutral", or "Positive"}
  \item{vidrep}{1,2,3}
  \item{mean_accuracy}{Mean accuracy}
  \item{mean_z_score}{Mean z-transformed RT}
}
}
\source{
Marianna
}
\usage{
data_mc
}
\description{
A data set with the RT scores of three groups.
}
\keyword{datasets}
