% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plot_my_var.R
\name{data_plot.my.var}
\alias{data_plot.my.var}
\title{data_plot.my.var}
\usage{
data_plot.my.var(
  data = tosave$DATA,
  var = NULL,
  var.pattern = NULL,
  var.exclude = NULL,
  var.order = NULL,
  id = NULL,
  equivalence.data = NULL,
  equivalence.var = NULL,
  equivalence.text = NULL,
  equivalence.gsub.patt = NULL,
  exclude = c(NA, "", " "),
  exclude.recod = "non-réponse",
  question.lab = NULL,
  var.group = NULL,
  var.group.order = c(),
  Rnd = 1
)
}
\arguments{
\item{data}{data.frame où sont stockées la.les variables(s)}

\item{var}{character. Nom de la variable}

\item{var.pattern}{character. Pattern qui identifie les variables.Pour variables multiples.}

\item{var.exclude}{character. Pattern a exclure de la recherche de variables. Pour variables multiples.}

\item{var.order}{vector of characters. Order des levels à que l'on souhaite spécifier.}

\item{equivalence.data}{data.frame donnant la structure du questionnaire, associant les noms de variables dans le data.frame original (equivalence.var) et un intitulé de question (equivalence.text).}

\item{exclude}{vector of characters. Levels à supprimer dans le décompte. Exemple : c(NA, "", " ", "non-réponse")}

\item{exclude.recod}{character. Si !is.null(exclude) permet de recoder les levels exclus. Exemple : "non-réponse".}

\item{question.lab}{Label de la question posée.}

\item{var.group}{faut-il dénombrer par groupes? Variable dans le data.frame original servat à grouper (character: nom de variable | numeric : position de la variable | un vecteur : la variable pour les groupes, de la même taille que la variable originale).}

\item{Rnd}{arrondi pour les pourcentages}
}
\description{
permet de générer un tableau de données qui compte les occurences des modalités d'une variable, associées à des méta-données. La variable peut être une colonne unique, ou une série de colonnes identifiées par un pattern (var.pattern).
}
