% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_tania}
\alias{data_tania}
\title{Stroop data}
\format{
A data frame:
\describe{
  \item{participant}{An identifier for the participant}
  \item{trial}{Trial number}
  \item{congruency}{ink color vs color word: "congruent", "incongruent"}
  \item{acc}{Accuracy for the current trial}
  \item{rt}{RT for the current trial}
}
}
\source{
Tania
}
\usage{
data_tania
}
\description{
A data set with RT and accuracy over many Stroop trials
}
\keyword{datasets}
