% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{Datos de Gapminder.}
\format{
El data frame \code{gapminder} contiene 1704 filas y 6 variables:
\describe{
\item{pais}{país al que corresponde la observación}
\item{continente}{categoría, continente donde se ubica el país del registro}
\item{anio}{("año") rango de 1952 a 2007 en incrementos de 5 años}
\item{expVida}{expectativa de vida al nacer, en años}
\item{pop}{población}
\item{PBIpc}{PBI per capita (USD, ajustados por inflación)}
}
}
\source{
\url{https://www.gapminder.org/}
}
\usage{
gapminder
}
\description{
Extracto de la base de datos de Gapminder con registros de expectativa de vida,
PBI per capita y población por país.
}
\keyword{datasets}
