% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plot_my_var.R
\name{gg.round_my_rounds}
\alias{gg.round_my_rounds}
\title{gg.round_my_rounds}
\usage{
gg.round_my_rounds(
  gg_obj = resd,
  ORDER = NULL,
  colpal = c("wes", "Darjeeling1"),
  sizepal = c(5, "PROP"),
  LENs = 30,
  lab.size = 3,
  family.L = "Ubuntu",
  NUDGE.X = 0.8,
  x.limit = c(-2.5, 2.5),
  y.limit = c(-1.1, 1.1),
  SIZE.Q = 5,
  family.Q = "Ubuntu Condensed",
  MOD.KEEP = NULL,
  ALPHA.RANGE = c(0.2, 1)
)
}
\arguments{
\item{gg_obj}{output de data_plot.my.var.}

\item{ORDER}{vector : labels présents dans les data, ordonnés : c("moda1", "moda2", ... )ze)}

\item{colpal}{list() ou vector(). colpal[1] donne le nom d'un paquet ("wes", "viridis" ou "manual") et colpal[2] donne le nom d'une palette (exemple : "Darjeeling1", "viridis" ou une palette manuelle).Si colpal[1]=="manual",  alors colpal[2] peut être un vector de couleurs nommé avec les levels présents dans les données, ou un character unique donnant une couleur unique pour tous les levels.}

\item{sizepal}{range pour les size des points. Si c(5, "PROP"), cela signifie que la taille min. est de 5 et qua la taille max est proportionnelle aux données. On peut spécifier aussi c("PROP", 10), ou c(1, 10).}

\item{lab.size}{size des labels}

\item{family.L}{famille de font des labels.}

\item{NUDGE.X}{espacement des labels avec les points}

\item{x.limit}{et y.limit : range du cadran pour le calcul du cercle (invisible, sur lequel sont positionnés les points) c(-2.5, 2.5)}

\item{y.limit}{et x.limit : range du cadran pour le calcul du cercle (invisible, sur lequel sont positionnés les points) c(-1.1, 1.1)}

\item{SIZE.Q}{taille du libellé de la question au centre du cercle}

\item{family.Q}{famille de fonts pour le libellé de la question au centre du cercle}

\item{MOD.KEEP}{NULL vecteur de modalités à conserver. Si spécifié, les autres modalités sont "effacées" avec alpha=0 (voir ALPHA.RANGE), mais leur position dans le cercle n'est pas réaffectée.}

\item{ALPHA.RANGE}{range d'alpha pour effacement des modalités (si MOD.KEEP!=NULL)}
}
\description{
permet de représenter des points en cercle: chaque point représente une modalité, avec un label, et une taille qui dépend du nombre d'occurences.
}
