% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plot_my_var.R
\name{gg_plot.my.var}
\alias{gg_plot.my.var}
\title{gg_plot.my.var}
\usage{
gg_plot.my.var(
  gg_obj = test,
  type = c("lollipop", "bar", "polar", "circumpolar", "bar*style"),
  orientation = "h",
  label.space = 0.1,
  label.space.x = 0,
  label.size.prop = 1,
  colpal = c("wes", "Darjeeling1"),
  base.size = 15
)
}
\arguments{
\item{gg_obj}{output de data_plot.my.var.}

\item{type}{c("lollipop", "bar", "polar", "circumpolar", "bar*style")}

\item{orientation}{"h" pour horizontal (par défaut).}

\item{label.space}{numérique. Espacement des labels avec le point (nudge).Ordonnée.}

\item{label.space.x}{numérique. Espacement des labels avec le point (nudge)Abcisse.}

\item{label.size.prop}{size of label in fraction of base.size (label.size.prop*base.size)}

\item{colpal}{list() ou vector(). colpal[1] donne le nom d'un paquet ("wes", "viridis" ou "manual") et colpal[2] donne le nom d'une palette (exemple : "Darjeeling1", "viridis" ou une palette manuelle).Si colpal[1]=="manual",  alors colpal[2] peut être un vector de couleurs nommé avec les levels présents dans les données, ou un character unique donnant une couleur unique pour tous les levels.}

\item{base.size}{Label de la question posée.}
}
\description{
permet de faire un ggplot() rapide à partir de l'output de data_plot.my.var.
}
\examples{
X11()
iris\%>\%data_plot.my.var(var = "Species", exclude = NULL, exclude.recod = NULL, question.lab = "Répartition des espèces")\%>\%gg_plot.my.var(type = "polar", label.size.prop = 0.35)
}
