% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plot_my_var_styles.R
\name{gg_plot_my_var_mystyle}
\alias{gg_plot_my_var_mystyle}
\title{gg_plot_my_var_mystyle}
\usage{
gg_plot_my_var_mystyle(
  gg_obj = test,
  colpal = c("wes", "Darjeeling1"),
  base.size = 15,
  label.size = 4.5,
  labels.n = TRUE,
  axis.lab.size = 10
)
}
\arguments{
\item{gg_obj}{voir gg_plot.my.var}

\item{colpal}{voir gg_plot.my.var}

\item{base.size}{voir gg_plot.my.var}

\item{label.size}{voir gg_plot.my.var}

\item{labels.n}{if TRUE : affiche les n() et les %. Si FALSE, que les %.}
}
\description{
voir gg_plot.my.var. Reprend gg_plot.my.var avec une pré-configuration MRIE (ne fonctionne que pour des geom_bar() ).
}
