% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ven.R
\name{gg_venn}
\alias{gg_venn}
\title{gg_venn}
\usage{
gg_venn(
  data,
  lev,
  HIGH_COL = gray(0.1),
  vars = c("Q2Q", "QP2", "Q1Q"),
  label.groups = c(as.character(atridata$question.text[atridata$names == "Q2Q"]),
    as.character(atridata$question.text[atridata$names == "QP2.MOUI."]),
    as.character(atridata$question.text[atridata$names == "Q1Q"])),
  label.groups.size = 2,
  label.count = TRUE,
  label.which = "both",
  label_alpha = 0.7,
  label.groups.alpha = 0.8,
  font_fam = "Ubuntu",
  font_face = "plain",
  PROP.COUNT.LABEL = TRUE,
  rnd = 1,
  add.exclude = NULL,
  exclude.pos = c(0.7, 0.8),
  add.rayon = 1,
  MARGS = margin(0, 0.5, 2, 0),
  sequential,
  ...
)
}
\arguments{
\item{data}{data.frame où sont stockées la.les variables(s)}

\item{lev}{modalité à aller chercher dans les variables pour construire les groupes, qui sera remplacée par TRUE.}

\item{HIGH_COL}{couleur de remplissage pour la zone le plus importante numériquement (pour la plus faible, blanc par défaut, puis un dégradé est calculé entre les deux).}

\item{vars}{variables concernées dans le df pour construire les groupes ; Exemple : c("Q2Q", "QP2", "Q1Q")}

\item{label.groups}{c("Label pour le groupe  vars[1]", "Label pour le groupe vars[2]", ... )}

\item{label.groups.size}{taille des labels des groupes}

\item{label.count}{Faut-il afficher les dénombement?}

\item{label.which}{si label.count==TRUE, quel type de décompte on affiche? "count", "percent" ou "both"}

\item{label_alpha}{alpha des labels de dénombrement}

\item{label.groups.alpha}{alpha des labels des noms de groupes}

\item{font_fam}{="Ubuntu"}

\item{font_face}{="plain",}

\item{sequential}{si TRUE, chaque groupe est représenté dans un plot seul, et un dernier plot rassemble tous les groupes. L'output est alors une list de plot.}
}
\description{
permet de générer un diagram de Venn avec 2 ou 3 modalités qui se croisent. Il s'agit d'une reprise du code proposé dans ce paquet : https://github.com/gaospecial/ggVennDiagram . Du paquet ggVennDiagram sont reprises les fonctions suivantes : st_multi_intersection, st_multi_difference, st_multi_union, multi_union, multi_intersect, multi_setdiff, three_dimension_circle_regions et two_dimension_circle_regions
}
