% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inALLboth.R
\name{\%inALLboth\%}
\alias{\%inALLboth\%}
\title{Is all y and x elements are commons, ignoring order?}
\usage{
x \%inALLboth\% y
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\description{
Is all y and x elements are commons, ignoring order?
}
