% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_knack.R
\name{knack.plot}
\alias{knack.plot}
\title{plot.knack}
\usage{
knack.plot(knat = catdes.123$Q1.Nb, col.neg = pickobrew(the.palette =
  "YlOrRd", the.indexes = 6), col.pos = pickobrew(the.palette = "YlGn",
  the.indexes = 4), col.na = gray(level = 0.95), ...)
}
\arguments{
\item{knat}{Aknat}

\item{col.neg}{col.neg}

\item{col.pos}{col.pos}

\item{col.na}{col.na}
}
\value{
a plot
}
\description{
plot.knack
}
