% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_a_cat.R
\name{map.a.cat}
\alias{map.a.cat}
\title{map.a.cat permet de tracer une carte avec des coloris dépendants d'une variable qualitative (factor).}
\usage{
map.a.cat(
  sp.object = france84,
  var = "datafactor",
  alternate.df = NULL,
  alternate.var = NULL,
  match.id = NULL,
  match.id.sp = NULL,
  qcol = col,
  borders = NULL,
  title = NULL,
  subtitle = NULL,
  all.poly = TRUE,
  all.poly.borders = NA
)
}
\arguments{
\item{sp.object}{A spatialpolygons or spatialpolygonsdataframe object}

\item{var}{Variable à maper du type sp.object@data$var. Pas une chaine de caractere.}

\item{alternate.df}{data.frame complémentaire si class(sp.object)=="SpatialPolygons"}

\item{alternate.var}{Variable à maper du type alternate.df$alternate.var Pas une chaine de caractere.}

\item{match.id.sp}{Variable dans le sp.object qui match les id de alternate.df. Chaine de caracteres.}

\item{qcol}{vecteur des couleurs. De meme longueur que qcut-1}

\item{borders}{Couleur des frontières. Vecteur, valeur unique ou NULL.}

\item{title}{Titre de la carte (par défaut: NULL)}

\item{subtitle}{Sous-titre de la  carte (par défaut: NULL)}

\item{all.poly}{FALSE. Fond de carte: faut-il tracer tous les polygones de l'objet sp.object?}

\item{all.poly.borders}{Frotières du fond de carte}

\item{match.id.df}{Variable dans le alternate.df qui match les id de sp.object. Chaine de caracteres.}
}
\value{
some value
}
\description{
map.a.cat permet de tracer une carte avec des coloris dépendants d'une variable qualitative (factor).
}
