% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_knack_internal.R
\name{plot.knack.internal}
\alias{plot.knack.internal}
\title{plot.knack.internal}
\usage{
\method{plot}{knack.internal}(
  kn = k,
  SEQ.CUT = seq.cut,
  col.neg = pickobrew(the.palette = "Reds", the.indexes = 8),
  col.pos = pickobrew(the.palette = "Greens", 8),
  col.na = pickobrew("Greys", 2),
  TITLE = "",
  FACET.VARS = facet.var,
  ...
)
}
\arguments{
\item{kn}{Aknat}

\item{col.neg}{col.neg}

\item{col.pos}{col.pos}

\item{col.na}{col.na}
}
\value{
a plot
}
\description{
plot.knack.internal
}
