% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temps.R
\docType{data}
\name{temps}
\alias{temps}
\title{Temperatura diaria en cuatro ciudades del mundo.}
\format{
un data frame con 1460 filas y 7 variables:
\describe{
\item{region}{región del mundo}
\item{country}{país donde se encuenta la ciudad}
\item{city}{ciudad donde se registró la temperatura}
\item{month}{mes del año}
\item{day_of_year}{día del año -de 1 a 365-}
\item{year}{año}
\item{carat}{temperatura promedio del día}
}
}
\source{
\url{https://www.kaggle.com/sudalairajkumar/daily-temperature-of-major-cities}
}
\usage{
temps
}
\description{
Dataset con registros diarios de la temperatura promedio en 2019 de de cuatro ciudades:
Beirut, Kiev, Tegucigalpa y Miami Beach.
}
\keyword{datasets}
