% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_forplot.R
\name{text.forplot}
\alias{text.forplot}
\title{text.forplot permet de changer des caracteres pour les rendres lisible par les polices 'Hershey'. Unicode anti-slash}
\usage{
\method{text}{forplot}(input = a)
}
\arguments{
\item{input}{une chaine de caracteres entrante}
}
\value{
input, une chaine de caracteres
}
\description{
text.forplot permet de changer des caracteres pour les rendres lisible par les polices 'Hershey'. Unicode anti-slash
}
