% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_MRIE_hc.R
\name{theme_MRIE_hc}
\alias{theme_MRIE_hc}
\title{theme_MRIE_hc}
\usage{
theme_MRIE_hc(
  base_size = 12,
  base_family = "sans",
  style = c("default", "darkunica"),
  coord_flip = FALSE,
  bgcolor = NULL
)
}
\arguments{
\item{base_size}{base text size}

\item{base_family}{font family}

\item{coord_flip}{les valeurs numériques sont-elles affichées en abscisse (x)?}
}
\description{
Thème pré-formaté MRIE pour ggplot. Reprend la structure des thèmes ggplot classiques et les arguments qui vont avec.
}
