% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_binaries.R
\name{transform_binaries}
\alias{transform_binaries}
\title{transform_binaries}
\usage{
transform_binaries(
  data = trest,
  Check_var = "Oui",
  No_var = "",
  type = "external",
  external.data = struct_jeune,
  external.data.text = "text",
  external.data.names = "names"
)
}
\arguments{
\item{data}{data.frame où sont stockées la.les variables(s)}
}
\description{
transforme une série de
}
