% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_it.R
\name{wrap.it}
\alias{wrap.it}
\title{wrap.it permet d'insérer des codes "\ n" (to the line) dans une chaine de caractere}
\usage{
wrap.it(x, len)
}
\arguments{
\item{x}{chaine de caractere}

\item{len}{length before wrapping}
}
\value{

}
\description{
wrap.it permet d'insérer des codes "\ n" (to the line) dans une chaine de caractere
}
