% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_labels.R
\name{wrap.labels}
\alias{wrap.labels}
\title{wrap.labels permet d'appliquer wrap.it à une liste ou un vecteur}
\usage{
wrap.labels(x, len)
}
\arguments{
\item{x}{chaine de caractere}

\item{len}{length before wrapping}
}
\description{
wrap.labels permet d'appliquer wrap.it à une liste ou un vecteur
}
