% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz-package.R
\docType{package}
\name{dataviz-package}
\alias{dataviz}
\alias{dataviz-package}
\title{dataviz: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
