% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_binaries.R
\name{transform_binaries}
\alias{transform_binaries}
\title{transform_binaries}
\usage{
transform_binaries(
  data = trest,
  Check_var = "Oui",
  No_var = "",
  type = "external",
  external.data = struct_jeune,
  external.data.text = "text",
  external.data.names = "names"
)
}
\arguments{
\item{data}{data.frame où sont stockées la.les variables(s)}

\item{var}{character. Nom de la variable}

\item{var.pattern}{character. Pattern qui identifie les variables.Pour variables multiples.}

\item{var.exclude}{character. Pattern a exclure de la recherche de variables. Pour variables multiples.}

\item{var.order}{vector of characters. Order des levels à que l'on souhaite spécifier.}

\item{exclude}{vector of characters. Levels à supprimer dans le décompte. Exemple : c(NA, "", " ", "non-réponse")}

\item{exclude.recod}{character. Si !is.null(exclude) permet de recoder les levels exclus. Exemple : "non-réponse".}

\item{question.lab}{Label de la question posée.}

\item{equivalence.data}{data.frame donnant la structure du questionnaire, associant les noms de variables dans le data.frame original (equivalence.var) et un intitulé de question (equivalence.text).}

\item{var.group}{faut-il dénombrer par groupes? Variable dans le data.frame original servat à grouper (character: nom de variable | numeric : position de la variable | un vecteur : la variable pour les groupes, de la même taille que la variable originale).}

\item{Rnd}{arrondi pour les pourcentages}
}
\description{
transforme une série de
}
